requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3